Backend2 0

+++ backend/src/controllers/documentos-trabajo.controller.ts
@@ -198,14 +198,40 @@ export class DocumentosTrabajoController {
   @ApiResponse({ status: 200, description: 'Elimina un documento por su ID.', type: DocumentoTrabajo })
   @ApiResponse({ status: 400, description: 'ID no válido' })
   @ApiResponse({ status: 500, description: 'Error interno del servidor' })
-  async delete(@Param('id') id: string): Promise<DocumentoTrabajo> {
+  async delete(
+    @Param('id') id: string,
+    @Headers('x-usuario-id') usuarioId: string
+  ): Promise<DocumentoTrabajo> {
     try {
       if (!Types.ObjectId.isValid(id)) {
         throw new BadRequestException('ID no válido');
       }
-      return this.documentosTrabajoService.delete(id);
+
+      if (!usuarioId) {
+        throw new BadRequestException('ID del usuario no proporcionado en el header x-usuario-id');
+      }
+
+      // Eliminar el documento
+      const documentoEliminado = await this.documentosTrabajoService.delete(id);
+      if (!documentoEliminado) {
+        throw new BadRequestException('Documento no encontrado');
+      }
+
+      // Registrar el log de la acción
+      const fecha = new Date();
+      await this.logsService.createLogDocument({
+        id_usuario: usuarioId,
+        id_documento: id,
+        accion: 'Eliminación documento',
+        fecha: fecha,
+      });
+
+      return documentoEliminado;
     } catch (error) {
-      console.error('Error al crear el libro:', error.message);
+      if (error instanceof BadRequestException) {
+        throw error;
+      }
+      console.error('Error al eliminar el documento:', error.message);
       throw new InternalServerErrorException('Error al eliminar el documento.');
     }
   }
